Refactor: pages 폴더 하위 section 분리
