+++ src/components/Navbar/index.js
@@ -1,1 +1,1 @@
-export { default } from "./Navbar.jsx";
+export { default } from "./Navbar";
