+++ src/pages/List/ChartSection.jsx
@@ -1,14 +1,11 @@
 import { useState } from "react";
-// Component
-import Button from "components/Button";
 import Tab from "components/Tab";
 import ChartItem from "components/ChartItem";
-// Design
+import Button from "components/Button";
 import { ReactComponent as ChartIcon } from "assets/icons/chart.svg";
 import femaleData from "mockData/femaleData";
 import maleData from "mockData/maleData";
 import styles from "./List.module.scss";
-// MockData
 
 export default function ChartSection() {
   const [sortedData, setSortedData] = useState(
