+++ src/styles/reset.scss
@@ -133,6 +133,8 @@ table {
   border-collapse: collapse;
 }
 
+/* Button Reset */
+
 button {
   padding: 0;
   border: none;
